requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
black==23.9.1
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.4
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
dnspython==2.3.0
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
fonttools==4.39.2
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.16.3
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
idna==3.4
importlib-metadata==6.1.0
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.4
libpff-python==20211114
loguru==0.6.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
numpy==1.24.1
openai==0.27.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pinecone-client==2.2.1
pipreqs==0.4.11
platformdirs==3.10.0
plotly==5.13.1
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.1.1
regex==2023.3.23
requests==2.28.2
rich==13.3.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.27.2
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.2
tokenize-rt==5.2.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
yarg==0.1.9
yarl==1.8.2
zipp==3.15.0
